Fix nterm and/or node locations whose right hand side start with actions
* expr_value_do

Node locations are not changed but `expr_value_do` locations are
fixed to not include spaces after `while`.

```
while a do end
```

* command_args

Node locations are not changed but `command_args` locations are
fixed to not include spaces after `cmd`.

```
cmd arg
```

* f_eq

Node locations are not changed but `f_eq` locations are
fixed to not include spaces after `arg`.

```
def m(arg = 1) end
```

* brace_body

* do_body

* f_arglist