Delay rb_parser_config_t free
Both parser and ast refer to rb_parser_config_t. The order of free of
these 2 struct is random because both are managed by GC.
Change to manage rb_parser_config_t with reference counter so that
rb_parser_config_t is freed when both parser and ast are freed.